+++ aws/alarm-investigation/extensions/models/aws/alarm_investigation_test.ts
@@ -3,8 +3,8 @@
 
 import { assertEquals, assertMatch, assertRejects } from "jsr:@std/assert@1";
 import { createModelTestContext } from "@systeminit/swamp-testing";
-import { CloudWatchClient } from "npm:@aws-sdk/client-cloudwatch@3.1069.0";
-import { SNSClient } from "npm:@aws-sdk/client-sns@3.1069.0";
+import { CloudWatchClient } from "npm:@aws-sdk/client-cloudwatch@3.1090.0";
+import { SNSClient } from "npm:@aws-sdk/client-sns@3.1090.0";
 import { model } from "./alarm_investigation.ts";
 
 // =============================================================================