+++ aws/networking/extensions/models/aws/networking.ts
@@ -13,17 +13,17 @@ import {
   DescribeAddressesCommand,
   DescribeNatGatewaysCommand,
   EC2Client,
-} from "npm:@aws-sdk/client-ec2@3.1069.0";
+} from "npm:@aws-sdk/client-ec2@3.1090.0";
 import {
   DescribeLoadBalancersCommand,
   DescribeTargetGroupsCommand,
   DescribeTargetHealthCommand,
   ElasticLoadBalancingV2Client,
-} from "npm:@aws-sdk/client-elastic-load-balancing-v2@3.1069.0";
+} from "npm:@aws-sdk/client-elastic-load-balancing-v2@3.1090.0";
 import {
   CloudWatchClient,
   GetMetricStatisticsCommand,
-} from "npm:@aws-sdk/client-cloudwatch@3.1069.0";
+} from "npm:@aws-sdk/client-cloudwatch@3.1090.0";
 
 const MAX_PAGES = 10;