+++ aws/service-quotas/extensions/models/aws/service-quotas.ts
@@ -17,21 +17,21 @@ import {
   ListServicesCommand,
   RequestServiceQuotaIncreaseCommand,
   ServiceQuotasClient,
-} from "npm:@aws-sdk/client-service-quotas@3.1069.0";
+} from "npm:@aws-sdk/client-service-quotas@3.1090.0";
 import {
   CloudWatchClient,
   GetMetricDataCommand,
-} from "npm:@aws-sdk/client-cloudwatch@3.1069.0";
+} from "npm:@aws-sdk/client-cloudwatch@3.1090.0";
 import {
   GetCallerIdentityCommand,
   STSClient,
-} from "npm:@aws-sdk/client-sts@3.1069.0";
-import { fromIni } from "npm:@aws-sdk/credential-providers@3.1069.0";
+} from "npm:@aws-sdk/client-sts@3.1090.0";
+import { fromIni } from "npm:@aws-sdk/credential-providers@3.1090.0";
 import {
   DescribeCasesCommand,
   DescribeCommunicationsCommand,
   SupportClient,
-} from "npm:@aws-sdk/client-support@3.1069.0";
+} from "npm:@aws-sdk/client-support@3.1090.0";
 
 // =============================================================================
 // Schemas